+++ src/components/Avatar/Avatar.example.js
@@ -1,6 +1,6 @@
 import Avatar, { AvatarMedium, AvatarLarge } from './Avatar';
 import { types as sdkTypes } from '../../util/sdkLoader';
-import { fakeIntl } from '../../util/test-data';
+import { fakeIntl } from '../../util/testData';
 
 const { UUID } = sdkTypes;