+++ src/components/ListingCard/ListingCard.test.js
@@ -1,12 +1,17 @@
 import React from 'react';
-import { renderDeep } from '../../util/test-helpers';
-import { createUser, createListing, fakeIntl } from '../../util/test-data';
+import '@testing-library/jest-dom';
+
+import { renderWithProviders as render } from '../../util/testHelpers';
+import { createUser, createListing, fakeIntl } from '../../util/testData';
+
 import { ListingCardComponent } from './ListingCard';
 
 describe('ListingCard', () => {
-  it('matches snapshot', () => {
+  test('matches snapshot', () => {
+    // This is quite small component what comes to rendered HTML
+    // For now, we rely on snapshot-testing.
     const listing = createListing('listing1', {}, { author: createUser('user1') });
-    const tree = renderDeep(<ListingCardComponent listing={listing} intl={fakeIntl} />);
-    expect(tree).toMatchSnapshot();
+    const tree = render(<ListingCardComponent listing={listing} intl={fakeIntl} />);
+    expect(tree.asFragment().firstChild).toMatchSnapshot();
   });
 });