+++ src/containers/AuthenticationPage/LoginForm/LoginForm.test.js
@@ -1,13 +1,30 @@
 import React from 'react';
-import { renderDeep } from '../../../util/test-helpers';
-import { fakeIntl } from '../../../util/test-data';
+import '@testing-library/jest-dom';
+
+import { renderWithProviders as render, testingLibrary } from '../../../util/testHelpers';
+import { fakeIntl } from '../../../util/testData';
+
 import LoginForm from './LoginForm';
+
+const { screen, userEvent } = testingLibrary;
 
 const noop = () => null;
 
 describe('LoginForm', () => {
-  it('matches snapshot', () => {
-    const tree = renderDeep(<LoginForm intl={fakeIntl} onSubmit={noop} />);
-    expect(tree).toMatchSnapshot();
+  test('Log in button is enabled when required fields are filled', () => {
+    render(<LoginForm intl={fakeIntl} onSubmit={noop} />);
+
+    // Test that sign up button is disabled at first
+    expect(screen.getByRole('button', { name: 'LoginForm.logIn' })).toBeDisabled();
+
+    // Type the values to the sign up form
+    userEvent.type(
+      screen.getByRole('textbox', { name: 'LoginForm.emailLabel' }),
+      'joe@example.com'
+    );
+    userEvent.type(screen.getByLabelText('LoginForm.passwordLabel'), 'secret-password');
+
+    // Test that sign up button is enabled after typing the values
+    expect(screen.getByRole('button', { name: 'LoginForm.logIn' })).toBeEnabled();
   });
 });